app/requirementes.txt
Flask==3.0.0
pymysql
python-dotenv==1.0.0
Werkzeug==3.0.1
